requirements.txt
bgg-api==1.1.10
python-telegram-bot==21.6
peewee==3.17.7

pytest==8.2.1
pytest-mock==3.14.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
ruff==0.4.8
mypy==1.10.0